+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export { FilterInput } from "./FilterInput";
-export { SelectionView } from "./SelectionView";
-export * from "./ShuffleButton";
+export { FilterInput } from "./Picker/FilterInput";
+export { SelectionView } from "./Picker/SelectionView";
+export * from "./Picker/ShuffleButton";
